+++ src/services/ghost/classic-auto-complete/__tests__/GhostContextProvider.test.ts
@@ -1,10 +1,9 @@
 import { describe, it, expect, beforeEach, vi } from "vitest"
-import { getProcessedSnippets, GhostContextProvider } from "../GhostContextProvider"
-import { AutocompleteInput } from "../../types"
+import { getProcessedSnippets } from "../GhostContextProvider"
+import { AutocompleteInput, GhostContextProvider } from "../../types"
 import { AutocompleteSnippetType } from "../../../continuedev/core/autocomplete/snippets/types"
 import { GhostModel } from "../../GhostModel"
 import { RooIgnoreController } from "../../../../core/ignore/RooIgnoreController"
-import * as vscode from "vscode"
 import crypto from "crypto"
 import { ContextRetrievalService } from "../../../continuedev/core/autocomplete/context/ContextRetrievalService"
 import { VsCodeIde } from "../../../continuedev/core/vscode-test-harness/src/VSCodeIde"
